@@ -1,6 +1,6 @@
 {
-  "name": "@rivercode/facebook-conversion-api-nextjs",
-  "version": "4.3.2",
+  "name": "@tapstack/facebook-conversion-api-nextjs",
+  "version": "5.2.0",
   "description": "Facebook Conversion API for Next.js",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -19,23 +19,24 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/RivercodeAB/facebook-conversion-api-nextjs.git"
+    "url": "git+https://github.com/tapstack/facebook-conversion-api-nextjs.git"
   },
-  "author": "Rivercode",
+  "author": "Tapstack",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/RivercodeAB/facebook-conversion-api-nextjs/issues"
+    "url": "https://github.com/tapstack/facebook-conversion-api-nextjs/issues"
   },
-  "homepage": "https://github.com/RivercodeAB/facebook-conversion-api-nextjs#readme",
+  "homepage": "https://github.com/tapstack/facebook-conversion-api-nextjs#readme",
   "dependencies": {
     "formdata-node": "^4.4.1",
     "universal-cookie": "^4.0.4",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "@tapstack/facebook-dash-conversion-api": "^1.0.0"
   },
   "peerDependencies": {
-    "next": "^11 || ^12 || ^13",
-    "react": "^16 || ^17 || ^18 || ^19",
-    "react-dom": "^16 || ^17 || ^18 || ^19"
+    "next": "^14 || ^15",
+    "react": "^18 || ^19",
+    "react-dom": "^18 || ^19"
   },
   "devDependencies": {
     "@types/node": "^18.16.1",
@@ -51,7 +52,7 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "next": "^13.3.1",
+    "next": "^15.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "^5.0.4"
