@@ -1,8 +1,5 @@
-const { createHash } = require('crypto');
+import { createHash } from 'crypto';
 
 const sha256Hash = (string: string) => createHash('sha256').update(string).digest('hex');
 
-export {
-  // eslint-disable-next-line import/prefer-default-export
-  sha256Hash,
-};
+export { sha256Hash };
